Separate single and multi brush tools and prototype box paint stamp
The new stamp multi brush uncovered a design flaw: since the sprite sheets are freeform, it's not easy to map them into discrete grid coordinates. This would be ok, if the map layer supports freeform shapes and this is the indended look, but for now, I actually want the grid snapping to be enforced.

Next, I want to figure out which information I could add in the SpriteEditor to the json files so mark a layer as grid-based or free-form.